ddddd

+++ controller/CartController.js
@@ -73,18 +73,31 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id;
+    let authorization = ensureAuthorization(req,res);
 
-    let sql = 'DELETE FROM cartItems WHERE id = ?';
-    conn.query(sql, cartItemId,
-        (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    if(authorization instanceof jwt.TokenExpiredError){
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
 
-        return res.status(StatusCodes.OK).json(results);
-    })
+    } else {
+        const cartItemId = req.params.id;
+
+        let sql = 'DELETE FROM cartItems WHERE id = ?';
+        conn.query(sql, cartItemId,
+            (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
 };
 
 module.exports = {
